server/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.112.1
h11==0.14.0
idna==3.7
Jinja2==3.1.4
livereload==2.7.0
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tornado==6.4.1
typing_extensions==4.12.2
uvicorn==0.30.6
